+++ src/Directory.Packages.props
@@ -15,11 +15,10 @@
     <PackageVersion Include="FluentValidation" Version="11.4.0" />
     <PackageVersion Include="GitHubActionsTestLogger" Version="2.0.1" />
     <PackageVersion Include="HdrHistogram" Version="2.5.0" />
-    <PackageVersion Include="Lucene.Net" Version="3.0.3" />
+    <PackageVersion Include="Lucene.Net.Analysis.Common" Version="4.8.0-beta00016" />
     <PackageVersion Include="Metrics.NET" Version="0.5.5" />
     <PackageVersion Include="Microsoft.AspNet.SignalR" Version="2.4.3" />
     <PackageVersion Include="Microsoft.AspNet.SignalR.Client" Version="2.4.3" />
-    <PackageVersion Include="Microsoft.AspNet.WebApi" Version="5.2.9" />
     <PackageVersion Include="Microsoft.AspNet.WebApi.OwinSelfHost" Version="5.2.9" />
     <PackageVersion Include="Microsoft.Data.SqlClient" Version="2.1.2" />
     <PackageVersion Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
@@ -84,9 +83,6 @@
   <ItemGroup Label="Versions to pin transitive references">
     <PackageVersion Include="Microsoft.Owin.Hosting" Version="4.2.2" />
     <PackageVersion Include="NServiceBus.Newtonsoft.Json" Version="2.4.0" />
-    <PackageVersion Include="System.Collections.Immutable" Version="7.0.0" />
-    <PackageVersion Include="System.Diagnostics.DiagnosticSource" Version="7.0.0" />
-    <PackageVersion Include="System.Memory" Version="4.5.5" />
     <PackageVersion Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
   </ItemGroup>
