+++ src/ServiceControl.Audit.Persistence.RavenDb/RavenDbAuditDataStore.cs
@@ -3,9 +3,6 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using System.Net;
-    using System.Net.Http;
-    using System.Net.Http.Headers;
     using System.Threading.Tasks;
     using Raven.Client;
     using ServiceControl.Audit.Auditing.MessagesView;
