+++ src/ServiceControl.Audit.Persistence.Tests.RavenDB/RavenPersistedTypes.cs
@@ -71,9 +71,9 @@ messageMetadata['ReceivingEndpoint']['$type'] = 'This.Is.The.Wrong.EndpointDetai
             string processingEndpoint = null
         )
         {
-            messageId = messageId ?? Guid.NewGuid().ToString();
-            conversationId = conversationId ?? Guid.NewGuid().ToString();
-            processingEndpoint = processingEndpoint ?? "SomeEndpoint";
+            messageId ??= Guid.NewGuid().ToString();
+            conversationId ??= Guid.NewGuid().ToString();
+            processingEndpoint ??= "SomeEndpoint";
 
             var metadata = new Dictionary<string, object>
             {
