+++ src/ServiceControl.Audit.Persistence.Tests/AuditTests.cs
@@ -231,10 +231,10 @@
             string messageType = null
         )
         {
-            messageId = messageId ?? Guid.NewGuid().ToString();
-            conversationId = conversationId ?? Guid.NewGuid().ToString();
-            processingEndpoint = processingEndpoint ?? "SomeEndpoint";
-            messageType = messageType ?? "MyMessageType";
+            messageId ??= Guid.NewGuid().ToString();
+            conversationId ??= Guid.NewGuid().ToString();
+            processingEndpoint ??= "SomeEndpoint";
+            messageType ??= "MyMessageType";
 
             var metadata = new Dictionary<string, object>
             {
