+++ src/ServiceControl.Audit.Persistence/Infrastructure/DeterministicGuid.cs
@@ -26,7 +26,7 @@
             Guid g;
 
             // use MD5 hash to get a 16-byte hash of the string
-            using (var provider = new MD5CryptoServiceProvider())
+            using (var provider = MD5.Create())
             {
                 var inputBytes = Encoding.Default.GetBytes(input);
                 var hashBytes = provider.ComputeHash(inputBytes);
