--- src/ServiceControl.Infrastructure.RavenDB/ServiceControl.Infrastructure.RavenDB.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="RavenDB.Database" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="$(PkgRavenDB_Database)\tools\Raven.Studio.Html5.zip" CopyToOutputDirectory="PreserveNewest" Visible="false" />
-  </ItemGroup>
-
-</Project>
