+++ src/ServiceControl.SagaAudit.Persistence.RavenDb/Extensions.cs
