+++ src/ServiceControl.Transports.ASBS/Helper.cs
@@ -23,7 +23,7 @@
             {
                 var inputBytes = Encoding.Default.GetBytes(input);
                 // use MD5 hash to get a 16-byte hash of the string
-                using (var provider = new MD5CryptoServiceProvider())
+                using (var provider = MD5.Create())
                 {
                     var hashBytes = provider.ComputeHash(inputBytes);
                     return new Guid(hashBytes).ToString();
