+++ src/ServiceControl.Transports.SqlServer/QueueLengthProvider.cs
@@ -5,9 +5,9 @@
     using System.Linq;
     using System.Threading.Tasks;
     using System.Collections.Concurrent;
-    using System.Data.SqlClient;
     using System.Threading;
     using NServiceBus.Logging;
+    using Microsoft.Data.SqlClient;
 
     class QueueLengthProvider : IProvideQueueLength
     {
@@ -27,7 +27,7 @@
 
             tableNames.AddOrUpdate(queueToTrack, _ => sqlTable, (_, currentSqlTable) =>
             {
-                if (currentSqlTable.Equals(sqlTable) == false)
+                if (!currentSqlTable.Equals(sqlTable))
                 {
                     tableSizes.TryRemove(currentSqlTable, out var _);
                 }
